feat: I set up state for the api data and the user, and make a get req.
I forgot to commit as I went along, but I have already finished the api calls and making the buttons to toggle between the data. Im using MUI to style my buttons and to make forms. Im also using a few hooks to keep track of state across the profile